fix controller and edit page

+++ backend/controller/MainTourTeamController.js
@@ -20,8 +20,9 @@ export const editMainTourTeam = async (req, res) => {
         team_id
     } = req.body;
     const {id} = req.params
-    if(!id && id === undefined) return res.status(401).json({message:'id is invalid'})
-
+    if (!id || isNaN(Number(id))) {
+        return res.status(401).json({ message: 'id is invalid' });
+    }
     const tour = prisma.t_tour_team.update({
         where: {
             id: Number(id),
@@ -33,11 +34,11 @@ export const editMainTourTeam = async (req, res) => {
     })
     res.json({status: 200, data: tour});
 }
-
 export const showMainTourTeam = async (req, res) => {
     const {id} = req.params
-    if(!id && id === undefined) return res.status(401).json({message:'id is invalid'})
-
+    if (!id || isNaN(Number(id))) {
+        return res.status(401).json({ message: 'id is invalid' });
+    }
     const tour = prisma.t_tour_team.findUnique({
         where: {
             id: Number(id),
@@ -46,7 +47,6 @@ export const showMainTourTeam = async (req, res) => {
 
     res.json({status: 200, data: tour});
 }
-
 export const showAllMainTourTeam = async (req, res) => {
     const tour = prisma.t_tour_team.findMany({
         include:{
@@ -72,7 +72,9 @@ export const deleteMainTourTeam = async (req, res) => {
 
 export const deleteAllMainTourTeam = async (req, res) => {
     const {id} = req.params
-
+    if (!id || isNaN(Number(id))) {
+        return res.status(401).json({ message: 'id is invalid' });
+    }
     const tour = prisma.t_tour_team.delete({
         where:{
             id: Number(id),
